requirements.txt
groq
python-dotenv
fpdf2
